stacks: add diagnostics to stack configuration
